High score menu finished
high score menu setup, a maximum amount of scores can be defined, but there's also a scroll bar, so maybe something up to a 100 before deleting would be good.

+++ Assets/_Scripts/DataHolder.cs
@@ -7,7 +7,7 @@ using System;
 using UnityEngine;
 
 public class DataHolder : IComparable<DataHolder>  {
-    public float ID { get; set; }
+    public int ID { get; set; }
 
     public float HighScore { get; set; }
 
@@ -17,7 +17,7 @@ public class DataHolder : IComparable<DataHolder>  {
 
     public float Skin2Unlock { get; set; }
 
-    public DataHolder(float id, string name, float highScore,   float skin1Unlock, float skin2Unlock)
+    public DataHolder(int id, string name, float highScore,   float skin1Unlock, float skin2Unlock)
     {
         this.HighScore = highScore;
         this.Name = name;
